Fixed ~/recipe/delete; delete is now performed by id (if there is a recipe with a corresponding id it will be deleted.

+++ cookbook/src/main/java/com/fooddude/cookbook/service/RecipeServiceImpl.java
@@ -36,7 +36,9 @@ public class RecipeServiceImpl implements RecipeService{
         return recipeRepository.save(recipe);
     }
     @Override
-    public void deleteRecipe(Recipe recipe) {
+    public void deleteRecipe(Integer id) throws InvalidRecipeIdException {
+        Recipe recipe = recipeRepository.findRecipeById(id, getAllRecipes());
+        if (recipe == null) throw new InvalidRecipeIdException(id);
         recipeRepository.delete(recipe);
     }
